# Avoid boxing in MultiVisitor