--- a/src/validation/mod.rs
+++ b/src/validation/mod.rs
@@ -14,7 +14,7 @@ pub use self::traits::Visitor;
 pub use self::visitor::visit;
 pub use self::context::{RuleError, ValidatorContext};
 pub use self::rules::visit_all_rules;
-pub use self::multi_visitor::MultiVisitor;
+pub use self::multi_visitor::{MultiVisitor, MultiVisitorNil};
 pub use self::input_value::validate_input_values;
 
 #[cfg(test)]