Get rid of all the boxing/heap allocations in `visit_all_rules`, and rely on a stack-allocated heterogenous linked list instead.

In my small GraphQL application (use-case is a high volume of rather simple queries), this brings me from about 17k to 20k rps. Working on getting some more detailed and relevant numbers, in the meantime I thought I'd already put this up for review.

Let me know what you think!