@srijs 2017-07-04
As promised, here are some more detailed numbers.

Without the patch:
```
Running 30s test @ http://localhost:3000/graphql
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   585.81us  405.66us  23.94ms   95.20%
    Req/Sec     8.51k     1.61k   12.04k    56.31%
  Latency Distribution
     50%  533.00us
     75%  692.00us
     90%  848.00us
     99%    1.23ms
  510122 requests in 30.10s, 70.54MB read
Requests/sec:  16947.29
Transfer/sec:      2.34MB
```

With the patch:
```
Running 30s test @ http://localhost:3000/graphql
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   537.27us  219.46us   8.27ms   75.96%
    Req/Sec     9.14k     1.89k   12.84k    53.41%
  Latency Distribution
     50%  493.00us
     75%  655.00us
     90%  806.00us
     99%    1.17ms
  546657 requests in 30.10s, 75.59MB read
Requests/sec:  18161.19
Transfer/sec:      2.51MB
```