@mhallin 2017-07-11
Nice, thanks!

As a comment on all these performance PRs; I think adding an LRU cache from Hash(query) to AST(query) for validated queries would be a great improvement for performance - parsing and validating is probably 90% of Juniper's overhead.

I've been thinking of implementing that, but it would require the API to expose some kind of global state object, or use `lazy_static!`. It might also require exposing some configuration parameters.